Final pylint fixes
